migrating to artemis ecs - basic rendering system - wip

+++ core/src/dev/manolovn/burger/components/Pos.kt
@@ -3,6 +3,6 @@ package dev.manolovn.burger.components
 import com.artemis.Component
 
 class Pos(
-        var x: Int = 0,
-        var y: Int = 0
+        var x: Float = 0f,
+        var y: Float = 0f
 ) : Component()
